refactor: 장르 저장 기능 분리
